requirements.txt
requests>=2.32.3
pillow>=10.3.0